extract styles from locationTracker and prediction screens
