+++ src/components/locationTracker/locationTracker.js
@@ -2,6 +2,7 @@ import React, { Component } from 'react';
 import { View, Image, StyleSheet, TextInput, TouchableOpacity, Text, AsyncStorage, StatusBar } from 'react-native';
 import DatePicker from 'react-native-datepicker'
 import sendLocation from '../../../src/fetch/createLocation'
+import styles from '../../../src/styles/locationTracker/locationTrackerStyles'
 
 export default class LocationTracker extends Component {
 
@@ -86,35 +87,3 @@ export default class LocationTracker extends Component {
     )
   }
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#0DA8AD'
-  },
-  info: {
-    color: '#FFF',
-    marginTop: 10,
-    opacity: 0.7,
-    fontSize: 20,
-    textAlign: 'center'
-  },
-  dateField: {
-    height: 200,
-    width: 200,
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  icon: {
-    height: 140,
-    width: 140,
-    alignItems: 'center',
-    justifyContent: 'center'
-  },
-  iconContainer: {
-    alignItems: 'center',
-    justifyContent: 'center',
-    height: 140,
-    width: 140,
-  }
-})
