+++ src/styles/locationTracker/locationTrackerStyles.js
@@ -1,0 +1,36 @@
+import React, { Component } from 'react';
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#0DA8AD'
+  },
+  info: {
+    color: '#FFF',
+    marginTop: 10,
+    opacity: 0.7,
+    fontSize: 20,
+    textAlign: 'center'
+  },
+  dateField: {
+    height: 200,
+    width: 200,
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+  icon: {
+    height: 140,
+    width: 140,
+    alignItems: 'center',
+    justifyContent: 'center'
+  },
+  iconContainer: {
+    alignItems: 'center',
+    justifyContent: 'center',
+    height: 140,
+    width: 140,
+  }
+})
+
+export default styles
