+++ src/styles/prediction/predictionStyles.js
@@ -1,0 +1,24 @@
+import React, { Component } from 'react';
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#0DA8AD'
+  },
+  info : {
+    color: '#FFF',
+    marginTop: 10,
+    opacity: 0.7,
+    fontSize: 20,
+    textAlign: 'center'
+  },
+  dateField: {
+    height: 200,
+    width: 200,
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+})
+
+export default styles
